updated folder streucture
