--- gateway/internal/middleware/logger.go
@@ -1,16 +1,0 @@
-package middleware
-
-import (
-	"github.com/alserov/restate/gateway/internal/log"
-	"github.com/alserov/restate/gateway/internal/wrappers"
-	"github.com/labstack/echo/v4"
-)
-
-func WithLogger(lg log.Logger) func(handlerFunc echo.HandlerFunc) echo.HandlerFunc {
-	return func(handlerFunc echo.HandlerFunc) echo.HandlerFunc {
-		return func(c echo.Context) error {
-			c.Set(string(wrappers.ContextLogger), lg)
-			return handlerFunc(c)
-		}
-	}
-}
