--- gateway/internal/wrappers/wrapper.go
@@ -1,8 +1,0 @@
-package wrappers
-
-type ContextKey string
-
-const (
-	ContextIdempotencyKey ContextKey = "key"
-	ContextLogger         ContextKey = "log"
-)
